Use short timeouts for HTTP requests
This fixes an issue where it was difficult to observe round robin load balancing
dynamics and intentions during training labs. The shorter 1 second timeout is
less optimal for a production scenario where we might want to use aggressive
caching, but it's better for a learning environment where we want to see the
effect of configuration changes.